+++ changelog.md
@@ -8,3 +8,4 @@
 - [FTBChunks] Improve chunk claiming mode
 - [FTBChunks] Improve client performance by changing how overlay label renders (A minimum version of JourneyMap 6.0.
   0.beta45 is required, or the labels won't render)
+- [FTBChunks] Fix overlay color does not update properly on team color changed